// UIAngular/ClientApp/src/app/core/backend/models/company-profile.ts
/* tslint:disable */
/* eslint-disable */
export interface CompanyProfile {
  about?: string;
  address?: string;
  coverImg?: string;
  followedHim?: boolean;
  followedMe?: boolean;
  followersCount?: number;
  followingCount?: number;
  fullName?: string;
  id?: string;
  interenId?: string;
  jobTitle?: string;
  mobile?: string;
  userImg?: string;
  userRole?: string;
}